WIP: RequiresWCSLinking plugin component mixin
* update alert language in orientation plugin
* footprints: if changing link type fails, redirect to orientation plugin - not ideal if the plugin is either inline or a popout... in the future we could detect whether there are subsets or markers and edit the text to say they will be removed (either in one button or additional buttons as-per the orientation plugin)
