+++ jdaviz/app.py
@@ -153,6 +153,7 @@ custom_components = {'j-tooltip': 'components/tooltip.vue',
                      'plugin-slider': 'components/plugin_slider.vue',
                      'plugin-color-picker': 'components/plugin_color_picker.vue',
                      'plugin-input-header': 'components/plugin_input_header.vue',
+                     'plugin-requires-wcs-linking': 'components/plugin_requires_wcs_linking.vue',
                      'glue-state-sync-wrapper': 'components/glue_state_sync_wrapper.vue',
                      'data-menu-add-data': 'components/data_menu_add_data.vue',
                      'data-menu-remove': 'components/data_menu_remove.vue',
